Update product assignment dialog messaages
