+++ src/orders/components/OrderProductAddDialog/styles.ts
@@ -55,5 +55,5 @@ export const useStyles = makeStyles(
       width: "100%",
     },
   }),
-  { name: "AssignVariantDialog" },
+  { name: "OrderProductAddDialog" },
 );
